requirements.txt
streamlit==1.52.2
pandas==2.3.3
plotly==6.5.0
requests==2.32.5
python-dotenv==1.2.1
pyarrow==22.0.0
koreanize-matplotlib==0.1.1
tabulate==0.9.0
matplotlib==3.10.8
numpy==2.4.0
